Moved files and packages model/view sources into separate directories
